Add transport layer
The Pocket API requires the consumer key and access token to be sent
with each request. So this would require passing those credentials
around, which seems a bad idea. Alternatively the credentials could be
set by an interceptor. This, however, would require constructing a JSON
object from the request, adding the key and token and then serialize
that object again. Quite some overhead. Using a transport object seems
to be the best choice.